Connector: Move models in file per connector
