+++ connectors/src/api/get_connector.ts
@@ -1,11 +1,8 @@
 import { Request, Response } from "express";
 
-import {
-  Connector,
-  GithubDiscussion,
-  GithubIssue,
-  NotionPage,
-} from "@connectors/lib/models";
+import { Connector } from "@connectors/lib/models";
+import { GithubDiscussion, GithubIssue } from "@connectors/lib/models/github";
+import { NotionPage } from "@connectors/lib/models/notion";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
 import { ConnectorType } from "@connectors/types/connector";
 import { ConnectorsAPIErrorResponse } from "@connectors/types/errors";
