+++ connectors/src/api/slack_channels_linked_with_agent.ts
@@ -7,7 +7,8 @@ import { Op } from "sequelize";
 import { joinChannel } from "@connectors/connectors/slack/lib/channels";
 import { getChannels } from "@connectors/connectors/slack/temporal/activities";
 import { APIErrorWithStatusCode } from "@connectors/lib/error";
-import { sequelize_conn, SlackChannel } from "@connectors/lib/models";
+import { sequelize_conn } from "@connectors/lib/models";
+import { SlackChannel } from "@connectors/lib/models/slack";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
 
 const PatchSlackChannelsLinkedWithAgentReqBodySchema = t.type({
