+++ connectors/src/api/webhooks/webhook_google_drive.ts
@@ -2,7 +2,7 @@ import { Request, Response } from "express";
 
 import { launchGoogleDriveIncrementalSyncWorkflow } from "@connectors/connectors/google_drive/temporal/client";
 import { APIErrorWithStatusCode } from "@connectors/lib/error";
-import { GoogleDriveWebhook } from "@connectors/lib/models";
+import { GoogleDriveWebhook } from "@connectors/lib/models/google_drive";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
 
 type GoogleDriveWebhookResBody = null | APIErrorWithStatusCode;
