+++ connectors/src/api/webhooks/webhook_slack.ts
@@ -8,11 +8,8 @@ import {
 } from "@connectors/connectors/slack/temporal/client";
 import { launchSlackGarbageCollectWorkflow } from "@connectors/connectors/slack/temporal/client";
 import { APIErrorWithStatusCode } from "@connectors/lib/error";
-import {
-  Connector,
-  SlackChannel,
-  SlackConfiguration,
-} from "@connectors/lib/models";
+import { Connector } from "@connectors/lib/models";
+import { SlackChannel, SlackConfiguration } from "@connectors/lib/models/slack";
 import { Ok } from "@connectors/lib/result";
 import logger from "@connectors/logger/logger";
 import { apiError, withLogging } from "@connectors/logger/withlogging";
