+++ connectors/src/connectors/notion/lib/connectors_db_helpers.ts
@@ -1,9 +1,5 @@
-import {
-  Connector,
-  ModelId,
-  NotionDatabase,
-  NotionPage,
-} from "@connectors/lib/models";
+import { Connector, ModelId } from "@connectors/lib/models";
+import { NotionDatabase, NotionPage } from "@connectors/lib/models/notion";
 import { DataSourceInfo } from "@connectors/types/data_source_config";
 
 // Note: this function does not let you "remove" a skipReason.
