+++ connectors/src/connectors/slack/lib/channels.ts
@@ -1,12 +1,8 @@
 import { CodedError, ErrorCode, WebAPIPlatformError } from "@slack/web-api";
 import { Channel } from "@slack/web-api/dist/response/ConversationsListResponse";
 
-import {
-  Connector,
-  ModelId,
-  sequelize_conn,
-  SlackChannel,
-} from "@connectors/lib/models";
+import { Connector, ModelId, sequelize_conn } from "@connectors/lib/models";
+import { SlackChannel } from "@connectors/lib/models/slack";
 import { Err, Ok, Result } from "@connectors/lib/result";
 import logger from "@connectors/logger/logger";
 import { ConnectorPermission } from "@connectors/types/resources";
